feat: extract theme controls module
